+++ framework/contracts/native/ans-host/src/tests/interact.rs
@@ -36,7 +36,7 @@ fn unauthorized_ans_host_update() {
     // Try adding a contract to the ans_host
     let msg = ExecuteMsg::UpdateContractAddresses {
         to_add: vec![(
-            "project:contract".to_string().try_into().unwrap(),
+            "project:contract".try_into().unwrap(),
             "contract_address".to_string(),
         )],
         to_remove: vec![],
@@ -78,7 +78,7 @@ fn authorized_ans_host_update() {
     // Try adding a contract to the ans_host
     let msg = ExecuteMsg::UpdateContractAddresses {
         to_add: vec![(
-            "project:contract".to_string().try_into().unwrap(),
+            "project:contract".try_into().unwrap(),
             "contract_address".to_string(),
         )],
         to_remove: vec![],
@@ -94,7 +94,7 @@ fn authorized_ans_host_update() {
     // Try removing a contract from the ans_host
     let msg = ExecuteMsg::UpdateContractAddresses {
         to_add: vec![],
-        to_remove: vec!["project:contract".to_string().try_into().unwrap()],
+        to_remove: vec!["project:contract".try_into().unwrap()],
     };
 
     let res = execute(deps.as_mut(), env, info, msg);