+++ integrations/astroport/contracts/pair_stable/tests/stablepool_tests.rs
@@ -185,8 +185,8 @@ fn swap_different_precisions() {
     let mut helper = Helper::new(&owner, test_coins.clone(), 100u64, None).unwrap();
 
     let assets = vec![
-        helper.assets[&test_coins[0]].with_balance(100_000_0000u128),
-        helper.assets[&test_coins[1]].with_balance(100_000_00000u128),
+        helper.assets[&test_coins[0]].with_balance(1_000_000_000_u128),
+        helper.assets[&test_coins[1]].with_balance(10_000_000_000_u128),
     ];
     helper.provide_liquidity(&owner, &assets).unwrap();
 
@@ -234,9 +234,9 @@ fn check_swaps() {
     let mut helper = Helper::new(&owner, test_coins.clone(), 100u64, None).unwrap();
 
     let assets = vec![
-        helper.assets[&test_coins[0]].with_balance(100_000_000000u128),
-        helper.assets[&test_coins[1]].with_balance(100_000_000000u128),
-        helper.assets[&test_coins[2]].with_balance(100_000_000000u128),
+        helper.assets[&test_coins[0]].with_balance(100_000_000_000_u128),
+        helper.assets[&test_coins[1]].with_balance(100_000_000_000_u128),
+        helper.assets[&test_coins[2]].with_balance(100_000_000_000_u128),
     ];
     helper.provide_liquidity(&owner, &assets).unwrap();
 
@@ -338,9 +338,9 @@ fn check_withdraw_charges_fees() {
     let mut helper = Helper::new(&owner, test_coins.clone(), 100u64, None).unwrap();
 
     let assets = vec![
-        helper.assets[&test_coins[0]].with_balance(100_000_000_000000u128),
-        helper.assets[&test_coins[1]].with_balance(100_000_000_000000u128),
-        helper.assets[&test_coins[2]].with_balance(100_000_000_000000u128),
+        helper.assets[&test_coins[0]].with_balance(100_000_000_000_000_u128),
+        helper.assets[&test_coins[1]].with_balance(100_000_000_000_000_u128),
+        helper.assets[&test_coins[2]].with_balance(100_000_000_000_000_u128),
     ];
     helper.provide_liquidity(&owner, &assets).unwrap();
 
@@ -416,7 +416,7 @@ fn check_pool_prices() {
                     .cumulative_prices
                     .iter()
                     .filter(|(from, to, _)| {
-                        from.eq(&helper.assets[&from_coin]) && to.eq(&helper.assets[&to_coin])
+                        from.eq(&helper.assets[from_coin]) && to.eq(&helper.assets[to_coin])
                     })
                     .collect::<Vec<_>>();
                 assert_eq!(price.len(), 1);
@@ -425,8 +425,8 @@ fn check_pool_prices() {
     };
 
     let assets = vec![
-        helper.assets[&test_coins[0]].with_balance(100_000_000_000000u128),
-        helper.assets[&test_coins[1]].with_balance(100_000_000_000000u128),
+        helper.assets[&test_coins[0]].with_balance(100_000_000_000_000_u128),
+        helper.assets[&test_coins[1]].with_balance(100_000_000_000_000_u128),
     ];
     helper.provide_liquidity(&owner, &assets).unwrap();
     helper.app.next_block(1000);
@@ -458,7 +458,7 @@ fn check_pool_prices() {
     helper.app.next_block(14 * 86400);
     check_prices(&helper);
 
-    let offer_asset = helper.assets[&test_coins[1]].with_balance(10_000_000000u128);
+    let offer_asset = helper.assets[&test_coins[1]].with_balance(10_000_000_000_u128);
     helper.give_me_money(&[offer_asset.clone()], &user1);
     helper
         .swap(