+++ integrations/astroport/contracts/tokenomics/xastro_token/src/testing.rs
@@ -615,7 +615,7 @@ fn snapshots_are_taken_and_retrieved_correctly() {
         current_block += 60_000;
 
         let transfer_amount = Uint128::new(10_000);
-        current_addr1_balance = current_addr1_balance - transfer_amount;
+        current_addr1_balance -= transfer_amount;
         current_addr2_balance += transfer_amount;
 
         let info = mock_info(addr1.as_str(), &[]);
@@ -640,8 +640,8 @@ fn snapshots_are_taken_and_retrieved_correctly() {
         current_block += 50_000;
 
         let burn_amount = Uint128::new(20_000);
-        current_total_supply = current_total_supply - burn_amount;
-        current_addr2_balance = current_addr2_balance - burn_amount;
+        current_total_supply -= burn_amount;
+        current_addr2_balance -= burn_amount;
 
         let info = mock_info(addr2.as_str(), &[]);